requirements.txt
# Development tools
pre-commit==4.2.0
pytest==7.4.0
pytest-cov==4.1.0

# Machine learning frameworks
pingouin==0.5.5
kagglehub==0.3.12
pycaret==3.3.2
sdv==1.21.0
mlflow==2.22.0
mlflow-skinny==2.22.0

# Core data science
numpy==1.26.4
pandas==2.1.4
matplotlib==3.7.5
seaborn==0.13.2
openpyxl==3.1.5
scipy==1.11.4
ucimlrepo==0.0.7
sklearn-compat==0.1.3
sklearn-pandas==2.2.0
statsmodels==0.14.4

# Deep learning
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
torchsummary==1.5.1

# Utilities
ipykernel==6.29.5
colorama==0.4.6